requirements.txt
# Google Cloud MLOps Project Dependencies
# Core ML Libraries
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Google Cloud Platform
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-cloud-core>=2.3.0

# Vertex AI Pipelines 
kfp>=2.4.0
google-cloud-pipeline-components>=2.5.0

# ML Training Frameworks
tensorflow>=2.15.0
joblib>=1.3.0

# Development Tools
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.27.0

# Data Processing
pyarrow>=14.0.0

# Configuration & Utils
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
